extract header provider to a separate file
